Update main.cpp
Tried to clear things up to setup how I want to try to work on this again. I plan to put more comments in here to further describe each bit of code I have. My main issues are the ordered list options and getting the variables to store inputs properly so the program can display the information as it was typed.